chore(sdk-coin-celo): move account-lib code
-moves code from account-lib to celo package

Ticket: BG-48146
