+++ modules/account-lib/package.json
@@ -37,13 +37,13 @@
     "@bitgo/sdk-coin-algo": "^1.0.0",
     "@bitgo/sdk-coin-avaxc": "^1.0.0",
     "@bitgo/sdk-coin-avaxp": "^1.0.0",
+    "@bitgo/sdk-coin-celo": "^1.0.0",
     "@bitgo/sdk-coin-cspr": "^1.0.0",
     "@bitgo/sdk-coin-eth": "^1.0.0",
     "@bitgo/sdk-coin-stx": "^1.0.0",
     "@bitgo/sdk-coin-xtz": "^1.0.0",
     "@bitgo/sdk-core": "^1.0.1",
     "@bitgo/statics": "^6.17.0",
-    "@celo/contractkit": "^2.0.0",
     "@ethereumjs/common": "^2.4.0",
     "@hashgraph/sdk": "~2.3.0",
     "@polkadot/api": "7.15.1",
@@ -62,24 +62,19 @@
     "bs58": "^4.0.1",
     "ethereumjs-abi": "^0.6.5",
     "ethereumjs-util": "6.2.1",
-    "ethers": "^5.1.3",
     "hi-base32": "^0.5.1",
     "joi": "^17.4.0",
     "js-sha256": "^0.9.0",
-    "keccak": "^3.0.2",
-    "libsodium-wrappers": "^0.7.6",
     "lodash": "^4.17.15",
     "long": "^4.0.0",
     "near-api-js": "^0.44.2",
-    "noble-bls12-381": "0.7.2",
     "protobufjs": "^6.8.9",
     "stellar-sdk": "^10.0.1",
     "tronweb": "^3.2.6",
     "tweetnacl": "^1.0.3"
   },
   "devDependencies": {
-    "@types/bs58": "^4.0.1",
-    "@types/libsodium-wrappers": "^0.7.7"
+    "@types/bs58": "^4.0.1"
   },
   "nyc": {
     "extension": [
