+++ modules/account-lib/tsconfig.json
@@ -20,6 +20,9 @@
       "path": "../sdk-coin-avaxp"
     },
     {
+      "path": "../sdk-coin-celo"
+    },
+    {
       "path": "../sdk-coin-cspr"
     },
     {
