+++ modules/bitgo/src/v2/coins/index.ts
@@ -16,6 +16,8 @@ import { Btc, Tbtc } from '@bitgo/sdk-coin-btc';
 export { Btc, Tbtc };
 import { Btg } from '@bitgo/sdk-coin-btg';
 export { Btg };
+import { Celo, CeloToken, Tcelo } from '@bitgo/sdk-coin-celo';
+export { Celo, CeloToken, Tcelo };
 import { Cspr, Tcspr } from '@bitgo/sdk-coin-cspr';
 export { Cspr, Tcspr };
 import { Dash, Tdash } from '@bitgo/sdk-coin-dash';
@@ -33,8 +35,6 @@ export { Xtz, Txtz };
 import { Zec, Tzec } from '@bitgo/sdk-coin-zec';
 export { Zec, Tzec };
 
-export { Celo } from './celo';
-export { CeloToken } from './celoToken';
 export { Eos } from './eos';
 export { EosToken } from './eosToken';
 export { Etc } from './etc';
@@ -46,7 +46,6 @@ export { StellarToken } from './stellarToken';
 export { Susd } from './susd';
 export { FiatUsd } from './fiatusd';
 export { FiatEur } from './fiateur';
-export { Tcelo } from './tcelo';
 export { Teos } from './teos';
 export { Tetc } from './tetc';
 export { Teth2 } from './teth2';
