+++ modules/sdk-coin-celo/package.json
@@ -42,8 +42,18 @@
   },
   "dependencies": {
     "@bitgo/abstract-eth": "^1.0.0",
+    "@bitgo/ethereumjs-utils-old": "npm:ethereumjs-util@5.2.0",
+    "@bitgo/sdk-coin-eth": "^1.0.0",
     "@bitgo/sdk-core": "^1.0.1",
-    "@bitgo/statics": "^6.17.0"
+    "@bitgo/statics": "^6.17.0",
+    "@celo/connect": "^2.0.0",
+    "@celo/contractkit": "^2.0.0",
+    "@celo/wallet-local": "^2.0.0",
+    "@ethereumjs/common": "^2.4.0",
+    "bignumber.js": "^9.0.0",
+    "ethereumjs-abi": "^0.6.5",
+    "ethereumjs-util": "6.2.1",
+    "ethers": "^5.1.3"
   },
   "devDependencies": {
     "@bitgo/sdk-api": "^1.0.1",
