+++ modules/sdk-coin-celo/src/celo.ts
@@ -4,7 +4,7 @@
 import { BaseCoin, BitGoBase } from '@bitgo/sdk-core';
 import { BaseCoin as StaticsBaseCoin } from '@bitgo/statics';
 import { AbstractEthLikeCoin } from '@bitgo/abstract-eth';
-import { Celo as CeloAccountLib } from '@bitgo/account-lib';
+import { KeyPair } from './lib';
 
 export class Celo extends AbstractEthLikeCoin {
   protected constructor(bitgo: BitGoBase, staticsCoin?: Readonly<StaticsBaseCoin>) {
@@ -18,7 +18,7 @@ export class Celo extends AbstractEthLikeCoin {
   isValidPub(pub: string): boolean {
     let valid = true;
     try {
-      new CeloAccountLib.KeyPair({ pub });
+      new KeyPair({ pub });
     } catch (e) {
       valid = false;
     }
