+++ modules/sdk-coin-celo/src/index.ts
@@ -1,3 +1,4 @@
 export * from './celo';
 export * from './tcelo';
 export * from './celoToken';
+export * from './lib';
