+++ modules/sdk-coin-celo/tsconfig.json
@@ -11,13 +11,22 @@
   "exclude": ["node_modules"],
   "references": [
     {
+      "path": "../abstract-eth"
+    },
+    {
       "path": "../sdk-api"
+    },
+    {
+      "path": "../sdk-coin-eth"
     },
     {
       "path": "../sdk-core"
     },
     {
       "path": "../sdk-test"
+    },
+    {
+      "path": "../statics"
     }
   ]
 }
